feat(routing): fibgroup removal from map
Fibgroup removal is tricky because fibgroups are shared by
fibroutes and mutable. We need to ensure that we don't remove
a fibgroup from the store if it is referenced by a fibroute.
If we did, the route would still point to the fibgroup, but
we would not be able to mutate it; hence the route would not
be updated if the fibgroup changed. On the other hand, we don't
want to leave stale fibgroups around.

Ideally, we would delete a fibgroup once the corresponding next-hop
was deleted. This patch adds a fib method to unregister a group,
called whenever a next-hop is destroyed. However, fibgroups are
not deleted if still have references to them. Therefore, as
a sanity, the purge method is still called. The reason for this is
to make sure we never delete a fibgroup that is referenced. This
should be easy to check, but there's a problem: because left-right
clones the underlaying data structure, refcounts get "doubled".
While this could be compensated changing the removal criteria
(remove if refcount reaches 2 instead of 1), for the time being,
the safest is to use the explicit unregister method introduced
in this patch, along with a call to purge() after fib route
deletions. By keeping the original removal criteria (which is
more restrictive) we ensure that we don't delete fibgroups
prematurely. By purging, we ensure that we will end-up deleting
those that are no longer in use, though not immediately.

Probably, the best approach may be to mark fibgroups as deletable
with the explicit unregister call when the corresponding next-hop
is removed, and delete those that get marked once no route referen-
ces them. However, this requires augmenting augmenting the
fibgroups structure with some flag, and the current implementation
works fine.

+++ routing/src/fib/fibtype.rs
@@ -101,27 +101,33 @@ impl Fib {
 
     /// Delete the [`FibRoute`] for a prefix
     fn del_fibroute(&mut self, prefix: Prefix) {
-        match prefix {
+        let removed = match prefix {
             Prefix::IPV4(p4) => {
                 if p4 == Ipv4Prefix::default() {
                     let route = FibRoute::with_fibgroup(self.groupstore.get_drop_fibgroup_ref());
-                    if let Some(_prior) = self.add_fibroute(Prefix::root_v4(), route) {
-                        // TODO: n.b. route increases refct of drop - that's ok
-                    }
+                    self.add_fibroute(Prefix::root_v4(), route)
                 } else {
-                    self.routesv4.remove(&p4);
+                    self.routesv4.remove(&p4)
                 }
             }
             Prefix::IPV6(p6) => {
                 if p6 == Ipv6Prefix::default() {
                     let route = FibRoute::with_fibgroup(self.groupstore.get_drop_fibgroup_ref());
-                    if let Some(_prior) = self.add_fibroute(Prefix::root_v6(), route) {}
+                    self.add_fibroute(Prefix::root_v6(), route)
                 } else {
-                    self.routesv6.remove(&p6);
+                    self.routesv6.remove(&p6)
                 }
             }
+        };
+        if removed.is_some() {
+            // here, we could iterate over the fibgroups of the removed route. However, in order to remove it
+            // from the group store, we'd need the key which we don't have. We could lookup the elements in the
+            // store matching each of the fibgroups (addresses) the route had, but it is simpler and probably
+            // faster to just purge. Since we still keep a ref to the removed route, let's make sure we drop
+            // it before we purge, so that it's references are gone before!
+            drop(removed);
+            self.groupstore.purge();
         }
-        self.groupstore.purge();
     }
 
     /// Set the [`Vtep`] for this [`Fib`]
@@ -252,8 +258,9 @@ impl Fib {
 }
 
 #[derive(Debug)]
-pub enum FibChange {
+enum FibChange {
     RegisterFibGroup((NhopKey, FibGroup)),
+    UnregisterFibGroup(NhopKey),
     AddFibRoute((Prefix, Vec<NhopKey>)),
     DelFibRoute(Prefix),
     SetVtep(Vtep),
@@ -264,6 +271,9 @@ impl Absorb<FibChange> for Fib {
         match change {
             FibChange::RegisterFibGroup((key, fibgroup)) => {
                 self.groupstore.add_mod_group(key, fibgroup.clone())
+            }
+            FibChange::UnregisterFibGroup(key) => {
+                self.groupstore.del(key);
             }
             FibChange::AddFibRoute((prefix, keys)) => self.build_add_fibroute(*prefix, keys),
             FibChange::DelFibRoute(prefix) => self.del_fibroute(*prefix),
@@ -294,6 +304,12 @@ impl FibWriter {
     pub fn register_fibgroup(&mut self, key: &NhopKey, fibgroup: &FibGroup, publish: bool) {
         self.0
             .append(FibChange::RegisterFibGroup((key.clone(), fibgroup.clone())));
+        if publish {
+            self.0.publish();
+        }
+    }
+    pub fn unregister_fibgroup(&mut self, key: &NhopKey, publish: bool) {
+        self.0.append(FibChange::UnregisterFibGroup(key.clone()));
         if publish {
             self.0.publish();
         }
